Added Mort's map stuff.
