+++ User Interface Expansion/MWSE/mods/UI Expansion/mcm.lua
@@ -202,7 +202,7 @@ function this.onCreate(container)
 
 	-- Credits:
 	mainPane:createLabel({ text = common.dictionary.configCredits }).borderTop = 6
-	mainPane:createLabel({ text = "  Programming: NullCascade, Hrnchamd, Petethegoat, Jiopsi, Remiros" })
+	mainPane:createLabel({ text = "  Programming: NullCascade, Hrnchamd, Petethegoat, Jiopsi, Remiros, Mort" })
 	mainPane:createLabel({ text = "  Colored Magic School Icons: R-Zero" })
 	mainPane:createLabel({ text = "  Inventory Filter Icons: Remiros" })
 	mainPane:createLabel({ text = "  Concepts and Testing: Morrowind Modding Community Discord" })
